fix: load service details by slug
