+++ src/app/services/[slug]/page.tsx
@@ -1,13 +1,20 @@
 import Link from 'next/link'
 import { headers } from 'next/headers'
 
-export default async function ServiceDetailsPage({ params }: { params: { id: string } }) {
-  const { id } = params
+export default async function ServiceDetailsPage({ params }: { params: { slug: string } }) {
+  const { slug } = params
   const baseUrl =
     process.env.NEXT_PUBLIC_BASE_URL || `http://${headers().get('host')}`
-  const res = await fetch(`${baseUrl}/api/services/${id}`, { cache: 'no-store' })
+  const res = await fetch(
+    `${baseUrl}/api/services/${encodeURIComponent(slug)}`,
+    { cache: 'no-store' }
+  )
   if (!res.ok) {
-    return <div className="text-red-500 text-xl p-8">Unable to load service details</div>
+    return (
+      <div className="text-red-500 text-xl p-8">
+        Unable to load service details
+      </div>
+    )
   }
   const service = await res.json()
